refactor: extract event-emitter to specific file (#1872)
