+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './iap';
 export * from './types';
+export * from './eventEmitter';
 export * from './hooks/useIAP';
 export * from './hooks/withIAPContext';
 export * from './purchaseError';
